These are the files for the interactive calendar
This is a basic interactive calendar. I made the javascript file so the calendar is interactive. As of now, the current day is highlighted, other days are highlighted when moused over, and the arrows can be used to change the month. New classes were made for the calendar in the MainStyle.css file. Some code was added to "MainPage.html" so "calendar.html" could be included. The layout can be changed if we want an event to be visible on the calendar or if we want to see the event when moused over. Bear with me, I am still new to using HTML, CSS, and JS so any changes or advice is welcome.