Add files via upload
You are now able to add events to the calendar. There is a pop-up object for creating an event and deleting an event. As of now, only one element cant be added to each date. The events are stored in local storage. Been attempting to allow multiple events on a date but having some issues. The events are there but they do not appear on the actual calendar. I'm still kinda new to Java Script so I am definitely just overlooking something. I will work on it and see if I can get multiple events to appear.